Add start of frame notification for lpcip3511
This enables the start of frame interrupt.  When a start of frame
interrupt occurs, a message is created and sent via
USB_DeviceNotificationTrigger.

Notifications are gated by the define USB_DEVICE_CONFIG_SOF_NOTIFICATIONS.
Some upper layers that use this middleware need start of frame
notifications for doing isochronous transfers.

This fixes the following issue:
https://github.com/nxp-mcuxpresso/mcux-sdk-middleware-usb/issues/7

Signed-off-by: James Zipperer <jzipperer@fb.com>