[MCUX-75039] fix cannot charge issue for usb_pd_charger_battery.
During ptn5110 initialization, set 0x60 to wrongly clear the
vbus monitor bit then cause vbus discharge can not be completed,
then the source role is blocked on waiting vbus safe0V.
The setting using 0x60 is used to fix another MCUX-65673
issue, but it seems that this issue is not fixed. Badly,
this change causes this new problem.

Signed-off-by: Aiden Hu <weiwei.hu@nxp.com>