[MCUX-75775] Fix overlapping dependencies issues on NPW.
This issue is caused by the dependency of middleware.usb.device.ehci,
that in some case the device need use driver.memory or
middleware.usb.phy or both. This makes it necessary in the case of
ecosystems to ensure that each subdependency is mutually exclusive and
that there is always one that can be satisfied.

Discard the overlapping dependencies to fix it.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>