[MCUX-80598][1/2] Fix auto-negotiation issue on ETH Adapter
This patch fix the issue that if macro EXAMPLE_PHY_INTERFACE_RGMII is
enabled, only 1Gbps speed ethernet environment can work, but cannot work
if link speed is 100Mbps or other speed environment.

The MII interface rate between the MAC and the PHY should be consistent
with the rate negotiated by the PHY, and if it is not, the MAC will not
work properly, that made this issue.

Split ETH_ADAPTER_HW_Init() into ETH_ADAPTER_PHY_Init() &
ETH_ADAPTER_MAC_Init(), and add new function ETH_ADAPTER_PHY_Reset().
ETH_ADAPTER_Init() is now used only for basic PHY and MAC
initialization, and ETH_ADAPTER_Reset() is used to configure the MAC if
a new link is established. Reconfiguration MAC when ethernet link is
changed to help compatibility with different ethernet environment rates.

Signed-off-by: Dv Alan <zhangyang.shen@nxp.com>